Spelling out each state of Automatic Package Restore
There's a lot of confusion of what the 'state' of the config file should look like under each version of Automatic Package Restore.  Based on the other documents, it's inferred what the state *should* be, but never outright explicitly stated.  This addition makes it clear which version produces which outcome.