Explain that vs templates should omit nuget refs
Explains more clearly that the project or item template authored into the VSIX should itself not include the artifacts that NuGet would add to a project. This seems non-obvious to folks since when they're reading these docs, they already have a working project template and they want to NuGet-tize it. So the first step is to actually _remove_ all the references from their template that they need, since NuGet will be responsible to add them right back. 
This commit adds the necessary explanatory text.