Fixed how-to enable nuget.exe mirror command
The instructions available now point to TeamCity build server, where, currently at least, Guests do not have access to any projects.

I found that the NuGet.ServerExtensions.dll is available on the CodePlex site for Nuget2. Also it appears that the dll is build against signed NuGet.exe (discussed in #559), so it is also needed to download the NuGet-Signed.exe.