Updated URLs to use the new SymbolSource service
As we've set up our new production symbol repository, I am proposing a change to the docs to reflect that. I will submit an issue to also update the hardcoded nuget.exe push URL.

Please note that our docs are not yet updated on the SymbolSource.org site, but we'd like all users to benefit from the faster and more stable engine as soon as possible.