changed "empty" version to indicate the lowest
According to this blog post (http://www.jeffreyfritz.com/2015/12/the-case-of-the-old-nuget-dependency-installation/) when a version is not specified the lowest version is used.  

"Starting with v2.8, NuGet is more conservative by default and will install the lowest version of packages that the package author depends on"

The docs state that the latest version is used which is not correct.