Edits to restore and cache topics
Restore condenses information from four topics into a single one that's
much more succint. Do give it lots of scrutiny to make sure nothing
important is omitted. In many cases the existing content looked outdata
or was redundant.

The cache troubleshooting topic I've renamed to Managing the NuGet Cache
and modified it to be a little more generic on the subject, given that
caching isn't discussed elsewhere.