Fixes/correction, restore TFBuild topic
Restoring the separate TFBuild topic (linked to from within Package
Restore). Fixed table in .NET Standard guide, fixes in packages and
source control.

This also includes an interim renaming of nuget-config-file-defaults to
nuget-config-defaults-file, which doesn't affect content, but I'd
already changed it in the VS project file so need to include it here.