Revision to main Create topic
The Creating a Package topic is reviewed, and pulls in content formerly
in Package Best Practices. A redirect is added for the best practices
topic; the .md file is renamed from Creat-a-Package to
Creating-a-Package so a redirect is added for that. Other files that
reference the create topic are updated to the correct link, including
those that used to refer to the separate best practices topic. Added one
image to the create topic to show a readme displayed.