Editing of Tools topics
Edited the four topics for the Package Manager UI, Console, nuget.exe,
and PowerShell commands. The Package Explorer topic is removed because
the content has been moved to
https://github.com/NuGetPackageExplorer/NuGetPackageExplorer. Links and
redirects have been updated accordingly, although I kept the graphics
here because there could still be extant links to them.

Note the change in redirect in consume\package-manager-ui.cshtml,
specifically to remove the extra anchor tag. This URL is what the
Package Manager UI in VS links to directly for help on the UI options,
and with this extra anchor those links would go to the wrong part of the
Package Manager UI topic. This change should correct that.

Also, I removed most of the graphics in the Package Manager Console
topic because they really didn't add much. It's better to simply point
to the reference where developers can see the full set of examples.